move userland plugin methods into bin, instead of it being a plugin
